A few changes
* Extract out MesosCallFactory
* Another pass on docs
* Connect sink and source with same kill switch
