+++ mesos-client/src/main/scala/mesosphere/mesos/client/StrictLoggingFlow.scala
@@ -1,0 +1,10 @@
+package mesosphere.mesos.client
+
+import akka.NotUsed
+import akka.stream.scaladsl.Flow
+import com.typesafe.scalalogging.StrictLogging
+
+// TODO - move somewhere shared
+trait StrictLoggingFlow extends StrictLogging {
+  protected def log[T](prefix: String): Flow[T, T, NotUsed] = Flow[T].map{ e => logger.info(s"$prefix$e"); e }
+}
